Распределяю константы и функции по модулям
