+++ js/util.js
@@ -36,10 +36,22 @@ const shuffleArray = function(array) {
   return array;
 };
 
+const generateWordsEnding = function (number, array) {
+  if (number === 1) {
+    return array[0];
+  } if (number > 1 && number < 5) {
+    return array [1];
+  }
+  if (number === 0 || number >=5) {
+    return array[2];
+  }
+}
+
 export {
   getRandomIntInclusive,
   getRandomFloatInclusive,
   generateRandomValue,
   generateRandomList,
-  shuffleArray
+  shuffleArray,
+  generateWordsEnding
 };
